"Add optional ignore for peerDependencies update
Introduced a new option isIgnoringPeerDependencies in the project builder. This option, when set to true, will bypass the updatePeerDependencies process providing more flexibility to the build process. This is a useful feature when dealing with scenarios where explicit control over dependency versioning is needed, or when updates are not immediately necessary or desirable."