"Refactor git status check in build.mjs
Updated the git status check to improve code readability and efficiency. Originally the status was only checked if it was not 'clean'. However, we now evaluate the status for both 'clean' and 'not clean' conditions thus eliminating duplicate checks. In addition, unnecessary semicolon at the end of the line was removed adhering to the proper JavaScript syntax rules for a more consistent, clean, and easy to understand code."