"Refactor project builder to prioritize library builds
Modified the project-builder.ts file to prioritize builds that are marked as 'isLibrary'. This is done by checking the 'isLibrary' property instead of 'isIgnoringBuild' and 'isIgnoringPeerDependencies'. This makes the project building process more streamlined for library projects by making sure only necessary steps are executed. Also, added a check in semver.ts to avoid unnecessary runs when the status is clean."