"Refactor build process, move semver operation
Optimised the building and versioning process by modifying the function in project-builder.ts. Previously, the process would error out if no publishDirectory was provided. Now, it merely skips the buildProject function but continues on to semver. Moreover, semver function was modified to handle absence of publishDirectory gracefully. This brings better flexibility, especially useful when semver increase process is skipped."