"Handle null publish directory in semver
Implemented a check to handle null values for publish directory in semver. If the value is provided, the directory will be used as a part of publishing path, otherwise the current directory will be used. This change prevents the code from breaking in case of null values and improves usability."